+++ src/main/kotlin/kutils/FileUtils.kt
@@ -1,0 +1,8 @@
+package kutils
+
+import java.io.File
+
+operator fun File.div(childName: String): File {
+    return this.resolve(childName)
+}
+